extract frequency chart code into its own component
